@@ -12,8 +12,8 @@ import 'package:flutter_tencentplayer_example/util/forbidshot_util.dart';
 
 class FullVideoPage extends StatefulWidget {
   PlayType playType;
-  String dataSource;
-  TencentPlayerController controller;
+  String? dataSource;
+  TencentPlayerController? controller;
 
   //UI
   bool showBottomWidget;
@@ -31,12 +31,12 @@ class FullVideoPage extends StatefulWidget {
 }
 
 class _FullVideoPageState extends State<FullVideoPage> {
-  TencentPlayerController controller;
-  VoidCallback listener;
+  TencentPlayerController? controller;
+  VoidCallback? listener;
 
   bool isLock = false;
   bool showCover = false;
-  Timer timer;
+  Timer? timer;
 
   _FullVideoPageState() {
     listener = () {
@@ -56,23 +56,23 @@ class _FullVideoPageState extends State<FullVideoPage> {
       DeviceOrientation.landscapeRight,
     ]);
     _initController();
-    controller.addListener(listener);
+    controller!.addListener(listener!);
     hideCover();
     ForbidShotUtil.initForbid(context);
     Screen.keepOn(true);
   }
 
   @override
-  Future dispose() {
+  void dispose() {
     super.dispose();
     SystemChrome.setEnabledSystemUIOverlays(
         [SystemUiOverlay.top, SystemUiOverlay.bottom]);
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
     ]);
-    controller.removeListener(listener);
+    controller!.removeListener(listener!);
     if (widget.controller == null) {
-      controller.dispose();
+      controller!.dispose();
     }
     ForbidShotUtil.disposeForbid();
     Screen.keepOn(false);
@@ -91,7 +91,7 @@ class _FullVideoPageState extends State<FullVideoPage> {
         controller = TencentPlayerController.asset(widget.dataSource);
         break;
       case PlayType.file:
-        controller = TencentPlayerController.file(widget.dataSource);
+        controller = TencentPlayerController.file(widget.dataSource!);
         break;
       case PlayType.fileId:
         controller = TencentPlayerController.network(null,
@@ -99,7 +99,7 @@ class _FullVideoPageState extends State<FullVideoPage> {
                 auth: {"appId": 1252463788, "fileId": widget.dataSource}));
         break;
     }
-    controller.initialize();
+    controller!.initialize();
   }
 
   @override
@@ -112,10 +112,10 @@ class _FullVideoPageState extends State<FullVideoPage> {
         },
         onDoubleTap: () {
           if (!widget.showBottomWidget || isLock) return;
-          if (controller.value.isPlaying) {
-            controller.pause();
+          if (controller!.value.isPlaying) {
+            controller!.pause();
           } else {
-            controller.play();
+            controller!.play();
           }
         },
         child: Container(
@@ -125,10 +125,10 @@ class _FullVideoPageState extends State<FullVideoPage> {
             alignment: Alignment.center,
             children: <Widget>[
               /// 视频
-              controller.value.initialized
+              controller!.value.initialized
                   ? AspectRatio(
-                      aspectRatio: controller.value.aspectRatio,
-                      child: TencentPlayer(controller),
+                      aspectRatio: controller!.value.aspectRatio,
+                      child: TencentPlayer(controller!),
                     )
                   : Image.asset('static/place_nodata.png'),
 
@@ -142,7 +142,7 @@ class _FullVideoPageState extends State<FullVideoPage> {
               Offstage(
                 offstage: isLock,
                 child: TencentPlayerGestureCover(
-                  controller: controller,
+                  controller: controller!,
                   showBottomWidget: widget.showBottomWidget,
                   behavingCallBack: delayHideCover,
                 ),
@@ -191,7 +191,7 @@ class _FullVideoPageState extends State<FullVideoPage> {
                           delayHideCover();
                           if (Platform.isAndroid) {
                             DeviceOrientation deviceOrientation =
-                                controller.value.orientation < 180
+                                controller!.value.orientation < 180
                                     ? DeviceOrientation.landscapeRight
                                     : DeviceOrientation.landscapeLeft;
                             if (isLock) {
@@ -258,16 +258,16 @@ class _FullVideoPageState extends State<FullVideoPage> {
     'http://1252463788.vod2.myqcloud.com/95576ef5vodtransgzp1252463788/e1ab85305285890781763144364/v.f30.mp4',
   ];
 
-  changeClear(int urlIndex, {int startTime}) {
-    controller?.removeListener(listener);
+  changeClear(int urlIndex, {int? startTime}) {
+    controller?.removeListener(listener!);
     controller?.pause();
     controller = TencentPlayerController.network(clearUrlList[urlIndex],
         playerConfig: PlayerConfig(
-            startTime: startTime ?? controller.value.position.inSeconds));
+            startTime: startTime ?? controller!.value.position.inSeconds));
     controller?.initialize().then((_) {
       if (mounted) setState(() {});
     });
-    controller?.addListener(listener);
+    controller?.addListener(listener!);
   }
 
   hideCover() {
@@ -280,7 +280,7 @@ class _FullVideoPageState extends State<FullVideoPage> {
 
   delayHideCover() {
     if (timer != null) {
-      timer.cancel();
+      timer?.cancel();
       timer = null;
     }
     if (showCover) {
