@@ -5,12 +5,12 @@ class CommonUtils {
 
   // 防抖函数: eg:输入框连续输入，用户停止操作300ms才执行访问接口
   static const deFaultDurationTime = 300;
-  static Timer timer;
+  static Timer? timer;
 
-  static antiShake(Function doSomething, {durationTime = deFaultDurationTime}) {
-    timer?.cancel();
+  static antiShake(Function? doSomething, {durationTime = deFaultDurationTime}) {
+    timer!.cancel();
     timer = new Timer(Duration(milliseconds: durationTime), () {
-      doSomething?.call();
+      doSomething!.call();
       timer = null;
     });
   }
@@ -18,10 +18,10 @@ class CommonUtils {
   // 节流函数: eg:300ms内，只会触发一次
   static int startTime = 0;
 
-  static throttle(Function doSomething, {durationTime = deFaultDurationTime}) {
+  static throttle(Function? doSomething, {durationTime = deFaultDurationTime}) {
     int currentTime = DateTime.now().millisecondsSinceEpoch;
     if (currentTime - startTime > durationTime) {
-      doSomething?.call();
+      doSomething!.call();
       startTime = DateTime.now().millisecondsSinceEpoch;
     }
   }
