@@ -4,15 +4,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_forbidshot/flutter_forbidshot.dart';
 
 class ForbidShotUtil {
-  static bool isCapture;
-  static StreamSubscription<void> subscription;
+  static bool? isCapture;
+  static StreamSubscription<void>? subscription;
 
   static initForbid(BuildContext context) async {
     isCapture = await FlutterForbidshot.iosIsCaptured;
     _deal(context);
     if (Platform.isIOS) {
       subscription = FlutterForbidshot.iosShotChange.listen((event) {
-        isCapture = !isCapture;
+        isCapture = !isCapture!;
         _deal(context);
       });
     }
