@@ -10,12 +10,12 @@ const List<double> rateList = [1.0, 1.2, 1.5, 2.0];
 
 class TencentPlayerBottomWidget extends StatefulWidget {
   final isShow;
-  final TencentPlayerController controller;
-  final VoidCallback behavingCallBack;
-  final ValueChanged<int> changeClear;
+  final TencentPlayerController? controller;
+  final VoidCallback? behavingCallBack;
+  final ValueChanged<int>? changeClear;
 
   // UI
-  final bool showClearBtn;
+  final bool? showClearBtn;
 
   TencentPlayerBottomWidget({this.isShow, this.controller, this.behavingCallBack, this.changeClear, this.showClearBtn});
 
@@ -24,7 +24,7 @@ class TencentPlayerBottomWidget extends StatefulWidget {
 }
 
 class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
-  TencentPlayerController get controller => widget.controller;
+  TencentPlayerController? get controller => widget.controller;
 
   int currentClearIndex = 0;
   bool isShowClearList = false;
@@ -62,17 +62,17 @@ class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
                   GestureDetector(
                     behavior: HitTestBehavior.opaque,
                     onTap: () {
-                      if (controller.value.isPlaying) {
-                        controller.pause();
+                      if (controller!.value.isPlaying) {
+                        controller!.pause();
                       } else {
-                        controller.play();
+                        controller!.play();
                       }
                       widget.behavingCallBack?.call();
                     },
                     child: Container(
                       height: _Style.bottomContainerH,
                       padding: EdgeInsets.all(15.0),
-                      child: Image.asset(controller.value.isPlaying ? 'static/player_pause.png' : 'static/player_play.png', width: _Style.iconPlayW, height: _Style.iconPlayW,),
+                      child: Image.asset(controller!.value.isPlaying ? 'static/player_pause.png' : 'static/player_play.png', width: _Style.iconPlayW, height: _Style.iconPlayW,),
                     ),
                   ),
                   /// 进度条
@@ -81,10 +81,10 @@ class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
                       behavingCallBack: () {
                         widget.behavingCallBack?.call();
                       },
-                      controller: controller,
+                      controller: controller!,
                     ),
                   ),
-                  widget.showClearBtn ? GestureDetector(
+                  widget.showClearBtn == true ? GestureDetector(
                     behavior: HitTestBehavior.opaque,
                     onTap: () {
                       setState(() {
@@ -116,7 +116,7 @@ class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
                         left: 15,
                         right: 20
                       ),
-                      child: Text('倍速${controller.value.rate}x', style: TextStyle(color: Colors.white, fontSize: 12,),),
+                      child: Text('倍速${controller!.value.rate}x', style: TextStyle(color: Colors.white, fontSize: 12,),),
                     ),
                   ),
                 ],
@@ -192,7 +192,7 @@ class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
                         behavior: HitTestBehavior.opaque,
                         onTap: () {
                           isShowRateList = false;
-                          controller.setRate(rate);
+                          controller!.setRate(rate);
                           widget.behavingCallBack?.call();
                           setState(() {
                           });
@@ -203,7 +203,7 @@ class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
                           decoration: BoxDecoration(
                               border: rate == rateList[rateList.length - 1] ? null : Border(bottom: BorderSide(width: 0.3, color: Color(0xffeeeeee)))
                           ),
-                          child: Text('$rate倍', style: controller.value.rate == rate ? TextStyle(color: Color(0xfff24724), fontSize: 12,) : TextStyle(color: Colors.white, fontSize: 12,),),
+                          child: Text('$rate倍', style: controller!.value.rate == rate ? TextStyle(color: Color(0xfff24724), fontSize: 12,) : TextStyle(color: Colors.white, fontSize: 12,),),
                         ),
                       );
                     }).toList(),
@@ -229,10 +229,10 @@ class _TencentPlayerBottomWidgetState extends State<TencentPlayerBottomWidget> {
 
 class BottomScrubber extends StatefulWidget {
   final TencentPlayerController controller;
-  final VoidCallback behavingCallBack; //正在交互
+  final VoidCallback? behavingCallBack; //正在交互
 
   BottomScrubber({
-    @required this.controller,
+    required this.controller,
     this.behavingCallBack,
   });
 
@@ -246,11 +246,11 @@ class _BottomScrubberState extends State<BottomScrubber> {
 
   TencentPlayerController get controller => widget.controller;
 
-  Duration seekPos;
+  Duration? seekPos;
 
   void seekToRelativePosition(Offset globalPosition) {
-    double width = WidgetUtil.findGlobalRect(currentKey).width;
-    double xOffSet = WidgetUtil.globalOffsetToLocal(currentKey, globalPosition).dx;
+    double width = WidgetUtil.findGlobalRect(currentKey)!.width;
+    double xOffSet = WidgetUtil.globalOffsetToLocal(currentKey, globalPosition)!.dx;
     final double relative = xOffSet / width;
     seekPos = controller.value.duration * relative;
     setState(() {
@@ -262,7 +262,7 @@ class _BottomScrubberState extends State<BottomScrubber> {
   }
   @override
   Widget build(BuildContext context) {
-    Duration showDuration = seekPos != null ? seekPos : controller.value.position;
+    Duration? showDuration = seekPos != null ? seekPos : controller.value.position;
 
     return Row(
       children: <Widget>[
@@ -284,7 +284,7 @@ class _BottomScrubberState extends State<BottomScrubber> {
                       backgroundColor: Color(0x33ffffff),
                     ),
                     TencentLinearProgressIndicator(
-                      value: controller.value.duration.inMilliseconds <= 0 ? 0 : showDuration.inMilliseconds / controller.value.duration.inMilliseconds,
+                      value: controller.value.duration.inMilliseconds <= 0 ? 0 : showDuration!.inMilliseconds / controller.value.duration.inMilliseconds,
                       valueColor: AlwaysStoppedAnimation<Color>(Color(0xfffe373c)),
                       backgroundColor: Colors.transparent,
                     ),
@@ -308,7 +308,7 @@ class _BottomScrubberState extends State<BottomScrubber> {
               seekToRelativePosition(details.globalPosition);
             },
             onHorizontalDragEnd: (DragEndDetails details) async {
-              await controller.seekTo(seekPos);
+              await controller.seekTo(seekPos!);
               seekPos = null;
               setState(() {
               });
@@ -321,7 +321,7 @@ class _BottomScrubberState extends State<BottomScrubber> {
                 return;
               }
               seekToRelativePosition(details.globalPosition);
-              await controller.seekTo(seekPos);
+              await controller.seekTo(seekPos!);
               seekPos = null;
               setState(() {
               });
@@ -335,7 +335,7 @@ class _BottomScrubberState extends State<BottomScrubber> {
           ),
           height: _Style.bottomContainerH,
           alignment: Alignment.centerRight,
-          child: Text(TimeUtil.formatDuration(showDuration) + '/' + TimeUtil.formatDuration(controller.value.duration), style: TextStyle(color: Colors.white, fontSize: 12,),),
+          child: Text(TimeUtil.formatDuration(showDuration!) + '/' + TimeUtil.formatDuration(controller.value.duration), style: TextStyle(color: Colors.white, fontSize: 12,),),
         ),
       ],
     );
