@@ -1,15 +1,15 @@
 class PlayerConfig {
   final bool autoPlay;
   final bool loop;
-  final Map<String, String> headers;
-  final String cachePath;
+  final Map<String, String>? headers;
+  final String? cachePath;
 
   // 单位:秒
   final double progressInterval;
 
   // 单位:秒
-  final int startTime;
-  final Map<String, dynamic> auth;
+  final int? startTime;
+  final Map<String, dynamic>? auth;
 
   // 后台播放
   final bool supportBackground;
