@@ -15,12 +15,12 @@ class TencentPlayer extends StatefulWidget {
 }
 
 class _TencentPlayerState extends State<TencentPlayer> {
-  VoidCallback _listener;
-  int _textureId;
+  VoidCallback? _listener;
+  int? _textureId = 0;
 
   _TencentPlayerState() {
     _listener = () {
-      final int newTextureId = widget.controller.textureId;
+      final int newTextureId = widget.controller.textureId!;
       if (newTextureId != _textureId) {
         setState(() {
           _textureId = newTextureId;
@@ -33,7 +33,7 @@ class _TencentPlayerState extends State<TencentPlayer> {
   void initState() {
     super.initState();
     _textureId = widget.controller.textureId;
-    widget.controller.addListener(_listener);
+    widget.controller.addListener(_listener!);
   }
 
   @override
@@ -42,19 +42,27 @@ class _TencentPlayerState extends State<TencentPlayer> {
     if (oldWidget.controller.dataSource != widget.controller.dataSource) {
       if (Platform.isAndroid) oldWidget.controller.dispose();
     }
-    oldWidget.controller.removeListener(_listener);
+    oldWidget.controller.removeListener(_listener!);
     _textureId = widget.controller.textureId;
-    widget.controller.addListener(_listener);
+    widget.controller.addListener(_listener!);
   }
 
   @override
   void deactivate() {
     super.deactivate();
-    widget.controller.removeListener(_listener);
+    widget.controller.removeListener(_listener!);
   }
 
   @override
   Widget build(BuildContext context) {
-    return _textureId == null ? Container() : RotatedBox(quarterTurns: (widget.controller.value.degree / 90).floor(), child: Texture(textureId: _textureId));
+    if (_textureId == null) {
+      return Container();
+    } else {
+      if ((widget.controller.value.degree / 90).floor() == 0) {
+        return Texture(textureId: _textureId!);
+      } else {
+        return RotatedBox(quarterTurns: (widget.controller.value.degree / 90).floor(), child: Texture(textureId: _textureId!));
+      }
+    }
   }
 }
