@@ -164,6 +164,9 @@ public class FlutterTencentplayerPlugin implements MethodCallHandler {
                 Map authMap = (Map<String, Object>)call.argument("auth");
                 authBuilder.setAppId(((Number)authMap.get("appId")).intValue());
                 authBuilder.setFileId(authMap.get("fileId").toString());
+                if (authMap.get("sign") != null) {
+                    authBuilder.setSign(authMap.get("sign").toString());
+                }
                 mVodPlayer.startPlay(authBuilder);
             } else {
                 // asset播放
@@ -194,13 +197,7 @@ public class FlutterTencentplayerPlugin implements MethodCallHandler {
                     // file、 network播放
                     String urlOrPath = call.argument("uri").toString();
 
-                    if (!urlOrPath.startsWith("http")) {
-                        // file
-                        MediaMetadataRetriever retr = new MediaMetadataRetriever();
-                        retr.setDataSource(urlOrPath);
-                        String rotation = retr.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_ROTATION);
-                        degree = Integer.parseInt(rotation);
-                    }
+                    degree = Util.getNetworkVideoRotate(urlOrPath);
 
                     mVodPlayer.startPlay(urlOrPath);
                 }
@@ -215,10 +212,14 @@ public class FlutterTencentplayerPlugin implements MethodCallHandler {
             switch (eventCode) {
                 //准备阶段
                 case TXLiveConstants.PLAY_EVT_VOD_PLAY_PREPARED:
+                    boolean isRotateAttri = degree == 90 || degree == 270;
+                    int width = isRotateAttri ? player.getHeight() : player.getWidth();
+                    int height = isRotateAttri ? player.getWidth() : player.getHeight();
+
                     playEventMap.put("event", "initialized");
                     playEventMap.put("duration", (int) player.getDuration());
-                    playEventMap.put("width", player.getWidth());
-                    playEventMap.put("height", player.getHeight());
+                    playEventMap.put("width", width);
+                    playEventMap.put("height", height);
                     playEventMap.put("degree", degree);
                     break;
                 case TXLiveConstants.PLAY_EVT_PLAY_PROGRESS:
@@ -257,7 +258,7 @@ public class FlutterTencentplayerPlugin implements MethodCallHandler {
             Map<String, Object> netStatusMap = new HashMap<>();
             netStatusMap.put("event", "netStatus");
             netStatusMap.put("netSpeed", param.getInt(TXLiveConstants.NET_STATUS_NET_SPEED));
-            netStatusMap.put("cacheSize", param.getInt(TXLiveConstants.NET_STATUS_V_SUM_CACHE_SIZE));
+            netStatusMap.put("fps", param.getInt(TXLiveConstants.NET_STATUS_VIDEO_FPS));
             eventSink.success(netStatusMap);
         }
 
